// src/app/core/features-control.types.ts
export enum AppIncompatabiliityReason {
  appDisabled = 'appDisabled',
  operatingSystemNotSupported = 'operatingSystemNotSupported',
  appVersionNotSupported = 'appVersionNotSupported',
  noInternet = 'noInternet',
  flagsmithUnreachable = 'flagsmithUnreachable',
  platformNotSupported = 'platformNotSupported'
};